+++ packages/utilities/projects/utilities/src/vanilla-extract/skeleton/index.ts
@@ -1,0 +1,1 @@
+export { skeletonAnimation, getSkeletonStyle } from './getSkeletonStyle';