+++ packages/utilities/projects/utilities/tests/unit/specs/jss-test.ts
@@ -1,10 +1,17 @@
+import { formElementLayeredGap } from '@porsche-design-system/components/src/styles/form-styles';
 import { type Styles, create } from 'jss';
 import jssPluginCamelCase from 'jss-plugin-camel-case';
 import jssPluginGlobal from 'jss-plugin-global';
 import jssPluginNested from 'jss-plugin-nested';
 import jssPluginSortMediaQueries from 'jss-plugin-sort-css-media-queries';
-import { getFocusStyle, getHoverStyle, getMediaQueryMax, getMediaQueryMin, getMediaQueryMinMax } from '../../../src/js';
-import { vanillaExtractGetMediaQueryMax } from './vanilla-extract-getMediaQueryMax.css';
+import {
+  getFocusStyle,
+  getHoverStyle,
+  getMediaQueryMax,
+  getMediaQueryMin,
+  getMediaQueryMinMax,
+  getSkeletonStyle,
+} from '../../../src/js';
 
 const jss = create({
   plugins: [