+++ packages/components/CHANGELOG.md
@@ -24,8 +24,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0), 
 
 #### Fixed
 
-- Styles: `vanilla-extract` returns `getMediaQueryMax`, `getMediaQueryMin` & `getMediaQueryMinMax` in wrong format
-  ([#3753](https://github.com/porsche-design-system/porsche-design-system/pull/3753))
+- Styles: `vanilla-extract` returns `getMediaQueryMax`, `getMediaQueryMin`, `getMediaQueryMinMax` & `getSkeletonStyle`
+  in wrong format ([#3753](https://github.com/porsche-design-system/porsche-design-system/pull/3753))
 
 ### [3.27.0-rc.0] - 2025-02-13